+++ internal/parquet/conversion_error.go
@@ -8,6 +8,7 @@ import (
 	"log/slog"
 	"os"
 	"path/filepath"
+	"strings"
 )
 
 // handleConversionError attempts to handle conversion errors by counting the number of lines in the file.
@@ -30,7 +31,7 @@ func handleConversionError(err error, path string) error {
 	slog.Error("parquet conversion failed", logArgs...)
 
 	// return wrapped error
-	return NewConversionError(fmt.Sprintf("failed to convert file: %s", err.Error()), rows, path)
+	return NewConversionError(err.Error(), rows, path)
 }
 
 func countLines(filename string) (int64, error) {
@@ -61,6 +62,7 @@ type ConversionError struct {
 	SourceFile   string
 	BaseError    error
 	RowsAffected int64
+	displayError string
 }
 
 func NewConversionError(msg string, rowsAffected int64, path string) *ConversionError {
@@ -68,9 +70,15 @@ func NewConversionError(msg string, rowsAffected int64, path string) *Conversion
 		SourceFile:   filepath.Base(path),
 		BaseError:    errors.New(msg),
 		RowsAffected: rowsAffected,
+		displayError: strings.Split(msg, "\n")[0],
 	}
 }
 
 func (c *ConversionError) Error() string {
-	return fmt.Sprintf("%s: %s", c.SourceFile, c.BaseError.Error())
+	return fmt.Sprintf("%s: %s", c.SourceFile, c.displayError)
+}
+
+// Merge adds a second error to the conversion error message.
+func (c *ConversionError) Merge(err error) {
+	c.BaseError = fmt.Errorf("%s\n%s", c.BaseError.Error(), err.Error())
 }